feat: add monotxconfig subcommand to generate a monorepo .tx/config file
The goal is to make it easier to create a .tx/config to contain all linked
resources so we can run statistics in a single place.

Be aware that the amount of resources is large and it will take a while
for downloading all resources form the platform via transifex cli.

Log: